Completed module 2

+++ module-2/challenge-1/main.go
@@ -3,6 +3,6 @@ package main
 import "fmt"
 
 func main() {
-	var someValue int = 999999999999999999999
+	var someValue int64 = 999999999999999999999
 	fmt.Println(someValue)
 }
